LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-justify/distribute-alias.tentative-expected.txt


FAIL text-justify: distribute is a parse-time alias of inter-character assert_equals: expected "inter-character" but got "distribute"
